clippy_lints: Enable empty_loop lint for no_std crates
We use a different message for no_std crates, so that we avoid
recomending functions the user cannot use. We also update the
documentation to note that the remediations are different for `std`
and `no_std` crates.

Furthermore, we move the `empty_loop` lint to a "Correctness" lint,
as writing `loop {}` can cause miscompilations w/ LLVM.

Signed-off-by: Joe Richey <joerichey@google.com>